GPP-495 popular api update

+++ geoss-proxy/src/main/java/com/eversis/esa/geoss/proxy/domain/PopularWord.java
@@ -20,6 +20,6 @@ public class PopularWord {
     /**
      * The Doc count.
      */
-    int docCount;
+    long docCount;
 
 }
